added changes by Oleksiy
